// src/app/modules/booking/booking.model.ts
import { Schema, model } from "mongoose";
import { TBooking } from "./booking.interface";

const bookingSchema = new Schema<TBooking>(
  {
    customer: { type: Schema.Types.ObjectId, ref: "User", required: true },
    service: { type: Schema.Types.ObjectId, ref: "Service", required: true },
    slot: { type: Schema.Types.ObjectId, ref: "Slot", required: true },
    vehicleType: {
      type: String,
      enum: [
        "car",
        "truck",
        "SUV",
        "van",
        "motorcycle",
        "bus",
        "electricVehicle",
        "hybridVehicle",
        "bicycle",
        "tractor",
      ],
      required: true,
    },
    vehicleBrand: { type: String, required: true },
    vehicleModel: { type: String, required: true },
    manufacturingYear: { type: String, required: true },
    registrationPlate: { type: String, required: true,unique:true },
    price: { type: Number, required: true },
    isPaid: { type: Boolean, default: false },
    transactionId: { type: String },
  },
  { timestamps: true }
);

const BookingModel = model<TBooking>("Booking", bookingSchema);
export default BookingModel;
